syscall: get rid of custom syscall implementation
Maintaining the custom system call code for each architecture
is a burden.  I'm not convinced that we have clobber registers
correct, and at best, we are just duplicating work that has
already been done by the glibc maintainers.

I asked Ben what the reasoning was behind not just using syscall.
This was his answer:

 The main issue was that glibc's pthreads implementation really
 sucked back during initial development and there was a use-case
 for having the io_XXX functions usable directly from clone()ed
 threads that didn't have all the glibc pthread state setup for
 per-cpu areas to handle per-thread errno.  That made sense back
 then, but is rather silly today.

This patch gets rid of the architecture specific system call
wrappers in favor of using syscall().  We leave the per-arch
header files with the syscall numbers because we'll be adding
a new system call soon, and this will allow us to add the
new number without requiring newer kernels on build servers.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>